adding desktop tracking hook
