+++ src/components/common/desctop-therapy-list/desctop-therapy-list.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { TherapyItem } from '@/components/common/therapy-item/therapy-item';
+import therapyData from '@/data/therapy.json';
+
+export const DesktopTherapyList = () => {
+  return (
+    <div className="xl:flex xl:flex-row justify-between">
+      {therapyData.therapySteps.map((item) => (
+        <TherapyItem key={item.id} item={item} />
+      ))}
+    </div>
+  );
+};
